pyproject.toml
[tool.poetry]
name = "seaport.py"
version = "1.0.0"
description = ""
authors = ["OpenSea Developers"]

[tool.poetry.dependencies]
python = "^3.9.12"
web3 = "^5.27.0"
pydantic = "^1.9.0"

[tool.poetry.dev-dependencies]
pyright = "1.1.235"
pylint = "2.13.4"
black = "22.1.0"
isort = "^5.10.1"
eth-brownie = "1.18.1"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
